+++ stac_fastapi/elasticsearch/stac_fastapi/elasticsearch/app.py
@@ -77,12 +77,14 @@ ENABLE_COLLECTIONS_SEARCH_ROUTE = get_bool_env(
     "ENABLE_COLLECTIONS_SEARCH_ROUTE", default=False
 )
 ENABLE_CATALOGS_ROUTE = get_bool_env("ENABLE_CATALOGS_ROUTE", default=False)
+HIDE_ALTERNATE_PARENTS = get_bool_env("HIDE_ALTERNATE_PARENTS", default=False)
 logger.info("TRANSACTIONS_EXTENSIONS is set to %s", TRANSACTIONS_EXTENSIONS)
 logger.info("ENABLE_COLLECTIONS_SEARCH is set to %s", ENABLE_COLLECTIONS_SEARCH)
 logger.info(
     "ENABLE_COLLECTIONS_SEARCH_ROUTE is set to %s", ENABLE_COLLECTIONS_SEARCH_ROUTE
 )
 logger.info("ENABLE_CATALOGS_ROUTE is set to %s", ENABLE_CATALOGS_ROUTE)
+logger.info("HIDE_ALTERNATE_PARENTS is set to %s", HIDE_ALTERNATE_PARENTS)
 
 settings = ElasticsearchSettings()
 session = Session.create_from_settings(settings)
@@ -231,6 +233,7 @@ if ENABLE_CATALOGS_ROUTE:
         catalogs_extension = CatalogsExtension(
             client=catalogs_client,
             settings=settings.model_dump(),
+            hide_alternate_parents=HIDE_ALTERNATE_PARENTS,
         )
         catalogs_transaction_extension = CatalogsTransactionExtension(
             client=catalogs_client,